+++ shell/client/notifications-client.js
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { computeTitleFromTokenOwnerUser } from "/imports/client/model-helpers.js";
+import { computeTitleFromTokenOwnerUser } from "/imports/model-helpers.js";
 
 testNotifications = () => {
   // Run on console to create some dummy notifications for the purpose of seeing what they look
